+++ backend/src/main/java/com/github/ryliecc/backend/models/categories/CategoryResponse.java
@@ -1,4 +1,4 @@
-package com.github.ryliecc.backend.models;
+package com.github.ryliecc.backend.models.categories;
 
 import lombok.Builder;
