+++ backend/src/main/java/com/github/ryliecc/backend/models/categories/TransactionCategory.java
@@ -1,10 +1,8 @@
-package com.github.ryliecc.backend.models;
+package com.github.ryliecc.backend.models.categories;
 
 import lombok.*;
 import org.springframework.data.mongodb.core.mapping.Document;
 import org.springframework.data.mongodb.core.mapping.MongoId;
-
-import java.time.Instant;
 
 @Document
 @NoArgsConstructor
@@ -12,12 +10,10 @@ import java.time.Instant;
 @Builder
 @Setter
 @Getter
-public class MonthlyRecurringTransaction {
+public class TransactionCategory {
     @MongoId
     private String id;
     private String title;
-    private Instant startDate;
-    private String amountOfMoney;
     private String creatorId;
-    private String transactionCategory;
+    private String categoryType;
 }
