+++ backend/src/main/java/com/github/ryliecc/backend/models/transaction/monthly/MonthlyRecurringTransaction.java
@@ -1,4 +1,4 @@
-package com.github.ryliecc.backend.models;
+package com.github.ryliecc.backend.models.transaction.monthly;
 
 import lombok.*;
 import org.springframework.data.mongodb.core.mapping.Document;
@@ -12,11 +12,11 @@ import java.time.Instant;
 @Builder
 @Setter
 @Getter
-public class TransactionEntry {
+public class MonthlyRecurringTransaction {
     @MongoId
     private String id;
     private String title;
-    private Instant timeLogged;
+    private Instant startDate;
     private String amountOfMoney;
     private String creatorId;
     private String transactionCategory;
