+++ backend/src/main/java/com/github/ryliecc/backend/service/CategoryRepo.java
@@ -1,6 +1,6 @@
 package com.github.ryliecc.backend.service;
 
-import com.github.ryliecc.backend.models.TransactionCategory;
+import com.github.ryliecc.backend.models.categories.TransactionCategory;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
