+++ backend/src/main/java/com/github/ryliecc/backend/service/MonthlyRecurringTransactionRepo.java
@@ -1,6 +1,6 @@
 package com.github.ryliecc.backend.service;
 
-import com.github.ryliecc.backend.models.MonthlyRecurringTransaction;
+import com.github.ryliecc.backend.models.transaction.monthly.MonthlyRecurringTransaction;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
