+++ backend/src/main/java/com/github/ryliecc/backend/service/TransactionRepo.java
@@ -1,6 +1,6 @@
 package com.github.ryliecc.backend.service;
 
-import com.github.ryliecc.backend.models.TransactionEntry;
+import com.github.ryliecc.backend.models.transaction.daily.TransactionEntry;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
